keyword-search/requirements.txt
Django==1.9.6
django-haystack==2.4.1
pytz==2016.4
whoosh==2.7.4
jieba==0.38
celery-haystack==0.10
django-celery
gunicorn==19.6.0
gevent==1.1.1
whitenoise==3.1
celery==3.1.24


